PDFUSION-main/BACKEND/requirements.txt
fastapi
Pillow
unstructured
Cython
tiktoken
unstructured[local-inference]
PyPDF2
mysql-connector-python
email-validator==1.1.2
fastapi==0.62.0
PyJWT==1.7.1
python-decouple==3.3
uvicorn==0.12.3
pydantic[email]
bcrypt